+++ CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [unreleased]
 
+- Adds UserMetadata recipe
+
 ## [0.5.5] - 2022-04-11
 ### Added 
 -   Adds functions for debug logging